requirements.txt
asgiref==3.3.1
Django==3.1.7
django-taggit==1.4.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
